Feasibility Study of a Mission to Remove Eight SL-8 Rocket Boosters from LEO
This paper presents a conceptual feasibility study of an active debris removal mission aiming to deorbit eight SL-8 Rocket Boosters from Low Earth Orbit (LEO). This is done by designing a satellite system consisting of eight CubeSatbased spacecraft, which will be deployed from a parking orbit. Each spacecraft will employ low thrust propulsion to approach a SL-8 booster, which will be captured and deorbited using a tethered net. The main subsystems of each spacecraft are conceptually designed. This includes power, structure, propulsion, communications and Attitude Determination and Control System (ADCS). The mission features in terms of required impulse and transfer times are computed, finding an ideal parking orbit for the mission. The launching phase is simulated in order to find the main parameters of a launcher vehicle capable of carrying the entire satellite system to the parking orbit. The sustainability aspects of the mission are analyzed. This includes an estimation of the cost, both for each subsystem and the launcher vehicle, yielding an estimated mission cost of less than USD 100M. The environmental impact of a mission of this kind is discussed, comparing the carbon dioxide emissions per rocket launch with a long-distance aeroplane flight. Furthermore, a short reflection on the social impact and importance of launching satellites is given. The paper concludes that the mission is not feasible at this moment in time due to the low technology readiness level of the propulsion subsystem.